Split typechecker call validation
